// src/app/models/successFactor.model.ts
export class responseSuccessFactor {
    documentType:  string;
    documentNumber:  string;
    fullName:  string;
    channel:  string;
    dateOfBirth:  string;
    company:  string;
    personalEmail:string;
    companyEmail:  string;
    region:  string;
    jobTitle:  string;
    jobCode:  string;
    subChannel: string;
    typeContract:  string;
    gender:  string;
    committeeArea:  string;
    addressArea:  string;
    management:  string;
    code:  string;
    description: string;
}